buttons functional, need completion

+++ src/main/java/rocks/zipcode/atm/CashMachine.java
@@ -44,7 +44,7 @@ public class CashMachine {
         );
     }
 
-    public void deposit(int amount) {
+    public void deposit(double amount) {
         if ((accountData != null) && (amount>0)) {
             tryCall(
                     () -> bank.deposit(accountData, amount),
@@ -53,7 +53,7 @@ public class CashMachine {
         }
     }
 
-    public void withdraw(int amount) {
+    public void withdraw(double amount) {
         if ((accountData != null) && (amount>0)) {
             tryCall(
                     () -> bank.withdraw(accountData, amount),
